requirements.txt
# Core dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.9.2

# Gemini AI
google-generativeai==0.8.3

# Image processing
Pillow==11.0.0

# CLI
click==8.1.7
tqdm==4.66.5

# Optional: for better async performance
aiofiles==24.1.0

# Configuration
pyyaml==6.0.2

